% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdal_grid.R
\name{gdal_grid}
\alias{gdal_grid}
\title{gdal_grid}
\usage{
gdal_grid(src_datasource, dst_filename, ot, of, txe, tye, outsize, a_srs,
  zfield, z_increase, z_multiply, a, spat, clipsrc, clipsrcsql, clipsrclayer,
  clipsrcwhere, l, where, sql, co, q, output_Raster = FALSE,
  ignore.full_scan = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{src_datasource}{Character. Any OGR supported readable datasource.}

\item{dst_filename}{Character. The GDAL supported output file.}

\item{ot}{Character. "type". For the output bands to be of the indicated data type.}

\item{of}{Character. "format". Select the output format. The default is GeoTIFF (GTiff). Use the short format name.}

\item{txe}{Numeric. c(xmin,xmax). Set georeferenced X extents of output file to be created.}

\item{tye}{Numeric. c(ymin,ymax). Set georeferenced Y extents of output file to be created.}

\item{outsize}{Numeric. c(xsize,ysize). Set the size of the output file in pixels and lines.}

\item{a_srs}{Character. "srs_def". Override the projection for the output file. The srs_def may be any of the usual GDAL/OGR forms, complete WKT, PROJ.4, EPSG:n or a file containing the WKT.}

\item{zfield}{Character. "field_name". Identifies an attribute field on the features to be used to get a Z value from. This value overrides Z value read from feature geometry record (naturally, if you have a Z value in geometry, otherwise you have no choice and should specify a field name containing Z value).}

\item{z_increase}{Numeric. increase_value. Addition to the attribute field on the features to be used to get a Z value from. The addition should be the same unit as Z value. The result value will be Z value + Z increase value. The default value is 0.}

\item{z_multiply}{Numeric. multiply_value. This is multiplication ratio for Z field. This can be used for shift from e.g. foot to meters or from elevation to deep. The result value will be (Z value + Z increase value) * Z multiply value. The default value is 1.}

\item{a}{Character. [algorithm[:parameter1=value1][:parameter2=value2]...] Set the interpolation algorithm or data metric name and (optionally) its parameters. See INTERPOLATION ALGORITHMS and DATA METRICS sections for further discussion of available options.}

\item{spat}{Numeric. c(xmin,ymin,xmax,ymax). Adds a spatial filter to select only features contained within the bounding box described by (xmin, ymin) - (xmax, ymax).}

\item{clipsrc}{Numeric or Character.  c(xmin,ymin,xmax,ymax)|WKT|datasource|spat_extent. Adds a spatial filter to select only features contained within the specified bounding box (expressed in source SRS), WKT geometry (POLYGON or MULTIPOLYGON), from a datasource or to the spatial extent of the -spat option if you use the spat_extent keyword. When specifying a datasource, you will generally want to use it in combination of the -clipsrclayer, -clipsrcwhere or -clipsrcsql options.}

\item{clipsrcsql}{Character. Select desired geometries using an SQL query instead.}

\item{clipsrclayer}{Character. "layername". Select the named layer from the source clip datasource.}

\item{clipsrcwhere}{Character. "expression". Restrict desired geometries based on attribute query.}

\item{l}{Character. "layername". Indicates the layer(s) from the datasource that will be used for input features. May be specified multiple times, but at least one layer name or a -sql option must be specified.}

\item{where}{Character. "expression". An optional SQL WHERE style query expression to be applied to select features to process from the input layer(s).}

\item{sql}{Character. "select_statement". An SQL statement to be evaluated against the datasource to produce a virtual layer of features to be processed.}

\item{co}{Character. "NAME=VALUE". Passes a creation option to the output format driver. Multiple -co options may be listed. See format specific documentation for legal creation options for each format.}

\item{q}{Logical. Suppress progress monitor and other non-error output.}

\item{output_Raster}{Logical. Return output dst_filename as a RasterBrick?}

\item{ignore.full_scan}{Logical. If FALSE, perform a brute-force scan if other installs are not found.  Default is TRUE.}

\item{verbose}{Logical. Enable verbose execution? Default is FALSE.}

\item{...}{Other parameters to pass to gdal_translate.}
}
\value{
NULL or if(output_Raster), a RasterBrick.
}
\description{
R wrapper for gdal_grid: creates regular grid from the scattered data
}
\details{
This is an R wrapper for the 'gdal_grid' function that is part of the
Geospatial Data Abstraction Library (GDAL).  It follows the parameter naming
conventions of the original function, with some modifications to allow for more R-like
parameters.  For all parameters, the user can use a single character string following,
precisely, the gdal_contour format (\url{http://www.gdal.org/gdal_grid.html}), or,
in some cases, can use R vectors to achieve the same end.

INTERPOLATION ALGORITHMS

There are number of interpolation algorithms to choose from.

invdist

Inverse distance to a power. This is default algorithm. It has following parameters:

power:
Weighting power (default 2.0).
smoothing:
Smoothing parameter (default 0.0).
radius1:
The first radius (X axis if rotation angle is 0) of search ellipse. Set this parameter to zero to use whole point array. Default is 0.0.
radius2:
The second radius (Y axis if rotation angle is 0) of search ellipse. Set this parameter to zero to use whole point array. Default is 0.0.
angle:
Angle of search ellipse rotation in degrees (counter clockwise, default 0.0).
max_points:
Maximum number of data points to use. Do not search for more points than this number. This is only used if search ellipse is set (both radii are non-zero). Zero means that all found points should be used. Default is 0.
min_points:
Minimum number of data points to use. If less amount of points found the grid node considered empty and will be filled with NODATA marker. This is only used if search ellipse is set (both radii are non-zero). Default is 0.
nodata:
NODATA marker to fill empty points (default 0.0).
invdistnn

(Since GDAL 2.1) Inverse distance to a power with nearest neighbor searching, ideal when max_points is used. It has following parameters:

power:
Weighting power (default 2.0).
radius:
The radius of the search circle, which should be non-zero. Default is 1.0.
max_points:
Maximum number of data points to use. Do not search for more points than this number. Found points will be ranked from nearest to furthest distance when weighting. Default is 12.
min_points:
Minimum number of data points to use. If less amount of points found the grid node is considered empty and will be filled with NODATA marker. Default is 0.
nodata:
NODATA marker to fill empty points (default 0.0).
average

Moving average algorithm. It has following parameters:

radius1:
The first radius (X axis if rotation angle is 0) of search ellipse. Set this parameter to zero to use whole point array. Default is 0.0.
radius2:
The second radius (Y axis if rotation angle is 0) of search ellipse. Set this parameter to zero to use whole point array. Default is 0.0.
angle:
Angle of search ellipse rotation in degrees (counter clockwise, default 0.0).
min_points:
Minimum number of data points to use. If less amount of points found the grid node considered empty and will be filled with NODATA marker. Default is 0.
nodata:
NODATA marker to fill empty points (default 0.0).
Note, that it is essential to set search ellipse for moving average method. It is a window that will be averaged when computing grid nodes values.

nearest

Nearest neighbor algorithm. It has following parameters:

radius1:
The first radius (X axis if rotation angle is 0) of search ellipse. Set this parameter to zero to use whole point array. Default is 0.0.
radius2:
The second radius (Y axis if rotation angle is 0) of search ellipse. Set this parameter to zero to use whole point array. Default is 0.0.
angle:
Angle of search ellipse rotation in degrees (counter clockwise, default 0.0).
nodata:
NODATA marker to fill empty points (default 0.0).
linear

(Since GDAL 2.1) Linear interpolation algorithm.

The Linear method performs linear interpolation by compution a Delaunay triangulation of the point cloud, finding in which triangle of the triangulation the point is, and by doing linear interpolation from its barycentric coordinates within the triangle. If the point is not in any triangle, depending on the radius, the algorithm will use the value of the nearest point or the nodata value.

It has following parameters:

radius:
In case the point to be interpolated does not fit into a triangle of the Delaunay triangulation, use that maximum distance to search a nearest neighbour, or use nodata otherwise. If set to -1, the search distance is infinite. If set to 0, nodata value will be always used. Default is -1.
nodata:
NODATA marker to fill empty points (default 0.0).
DATA METRICS

Besides the interpolation functionality gdal_grid can be used to compute some data metrics using the specified window and output grid geometry. These metrics are:

minimum:
Minimum value found in grid node search ellipse.

maximum:
Maximum value found in grid node search ellipse.

range:
A difference between the minimum and maximum values found in grid node search ellipse.

count:
A number of data points found in grid node search ellipse.

average_distance:
An average distance between the grid node (center of the search ellipse) and all of the data points found in grid node search ellipse.

average_distance_pts:
An average distance between the data points found in grid node search ellipse. The distance between each pair of points within ellipse is calculated and average of all distances is set as a grid node value.

All the metrics have the same set of options:

radius1:
The first radius (X axis if rotation angle is 0) of search ellipse. Set this parameter to zero to use whole point array. Default is 0.0.
radius2:
The second radius (Y axis if rotation angle is 0) of search ellipse. Set this parameter to zero to use whole point array. Default is 0.0.
angle:
Angle of search ellipse rotation in degrees (counter clockwise, default 0.0).
min_points:
Minimum number of data points to use. If less amount of points found the grid node considered empty and will be filled with NODATA marker. This is only used if search ellipse is set (both radii are non-zero). Default is 0.
nodata:
NODATA marker to fill empty points (default 0.0).


This function assumes the user has a working GDAL on their system.  If the
"gdalUtils_gdalPath" option has been set (usually by gdal_setInstallation),
the GDAL found in that path will be used.  If nothing is found, gdal_setInstallation
will be executed to attempt to find a working GDAL that has the right drivers
as specified with the "of" (output format) parameter.

The user can choose to (optionally) return a RasterBrick of the output file (assuming
raster/rgdal supports the particular output format).
}
\examples{
# We'll pre-check to make sure there is a valid GDAL install
# and that raster and rgdal are also installed.
# Note this isn't strictly neccessary, as executing the function will
# force a search for a valid GDAL install.
gdal_setInstallation()
valid_install <- !is.null(getOption("gdalUtils_gdalPath"))
if(require(raster) && require(rgdal) && valid_install)
{
# Example from the original gdal_contour documentation:
# 	gdal_contour -a elev dem.tif contour.shp -i 10.0
# Choose a DEM:
input_dem  <- system.file("external/tahoe_lidar_bareearth.tif", package="gdalUtils")
# Setup an output filename (shapefile):
output_shapefile <- paste(tempfile(),".shp",sep="")
contour_output <- gdal_contour(src_filename=input_dem,dst_filename=output_shapefile,
		a="Elevation",i=5.,output_Vector=TRUE)
# Plot the contours using spplot:
spplot(contour_output["Elevation"],contour=TRUE)
}
}
\author{
Jonathan A. Greenberg (\email{gdalUtils@estarcion.net}) (wrapper) and Frank Warmerdam (GDAL lead developer).
}
\references{
\url{http://www.gdal.org/gdal_grid.html}
}

