% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.R
\name{qm}
\alias{qm}
\title{qm}
\usage{
qm(x)
}
\arguments{
\item{x}{Character or Numeric.}
}
\value{
A character string that begins and ends with quotation marks.
}
\description{
Wraps an input in quotation marks.
}
\examples{
{
qm("Hi!")
qm(42) 
}
}
\author{
Jonathan A. Greenberg
}
