% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_file_extension.R
\name{remove_file_extension}
\alias{remove_file_extension}
\title{remove_file_extension}
\usage{
remove_file_extension(filename, extension_delimiter = ".")
}
\arguments{
\item{filename}{Character. The input filename.}

\item{extension_delimiter}{Character. The extension or extension delimiter (default ".") to remove.}
}
\description{
Strips a file extension from a filename.
}
\examples{
myfilename="my.file.gri"
remove_file_extension(myfilename,".")
remove_file_extension(myfilename,".file.gri")
}
\author{
Jonathan A. Greenberg \email{spatial.tools@estarcion.net}
}
