% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_projlibpath.R
\name{set_projlibpath}
\alias{set_projlibpath}
\title{set_projlibpath}
\usage{
set_projlibpath()
}
\value{
The function is used for its side effects
}
\description{
Function needed to avoid problems related to side effects upon loading
 of rgdal/sf on Windows systems with GDAL 3 installed (see https://github.com/r-spatial/discuss/issues/31)
}
\details{
The function automatically resets the PROJ_LIB environment variable to the 
 "correct" folder for a GDAL3 installation, and restores the previous value upon 
 exiting from the caller function.
}
