% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tahoe_lidar_bareearth_data.R
\docType{data}
\name{tahoe_lidar_bareearth.tif}
\alias{tahoe_lidar_bareearth.tif}
\title{Lidar-derived bare earth digital elevation model from the Lake Tahoe Basin.}
\description{
Lidar-derived bare earth digital elevation model from the Lake Tahoe Basin.
}
\examples{
\dontrun{
tahoe_lidar_bareearth <- 
raster(system.file("external/tahoe_lidar_bareearth.tif", package="gdalUtils"))
plot(tahoe_lidar_bareearth)
}
}
\author{
Jonathan A. Greenberg \email{gdalUtils@estarcion.net}
}
\keyword{data}
