% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{test_modis.hdf}
\alias{test_modis.hdf}
\title{MODIS HDF4 file}
\description{
MODIS HDF4 file
}
\examples{
\dontrun{
gdalinfo(system.file("external/test_modis.hdf", package="gdalUtils"))
}
}
\author{
Jonathan A. Greenberg \email{gdalUtils@estarcion.net}
}
\keyword{data}

